binder, epoch, registry: minor refactor and example explicit nonce
